Simplify audit activities for user deletion
Instead of creating an user deletion activity and then registering
a further activity for the assignee update, we will just set a
unique activity for the investigation being
"automatically reassigned".
