+++ psd-web/app/models/audit_activity/investigation/update_assignee.rb
@@ -1,7 +1,25 @@
-class AuditActivity::Investigation::UpdateAssignee::WithNotification < AuditActivity::Investigation::UpdateAssignee::Base
+class AuditActivity::Investigation::UpdateAssignee < AuditActivity::Investigation::Base
+  include NotifyHelper
+
   def self.from(investigation)
+    title = investigation.assignee.id.to_s
     body = investigation.assignee_rationale
-    super(investigation, body)
+    super(investigation, title, self.sanitize_text(body))
+  end
+
+  def subtitle_slug
+    "Assigned"
+  end
+
+  def assignable_id
+    # We store assignable_id in title field, this is getting it back
+    # Using alias for accessing parent method causes errors elsewhere :(
+    AuditActivity::Investigation::Base.instance_method(:title).bind(self).call
+  end
+
+  def title
+    # We store assignable_id in title field, this is computing title based on that
+    "Assigned to #{(User.find_by(id: assignable_id) || Team.find_by(id: assignable_id))&.display_name}"
   end
 
   def email_update_text
