--- psd-web/app/models/audit_activity/investigation/update_assignee/without_notification.rb
@@ -1,9 +1,0 @@
-class AuditActivity::Investigation::UpdateAssignee::WithoutNotification < AuditActivity::Investigation::UpdateAssignee::Base
-  def entities_to_notify
-    []
-  end
-
-  def email_subject_text; end
-
-  def email_update_text; end
-end
